# Add direct conversion from spec.Schema to SMD schema.Schema