… pkg/schemaconv/openapi_test.go
… pkg/schemaconv/testdata/crds/openapiv2/accesscontextmanager_v1beta1_accesscontextmanagerserviceperimeter.json
… pkg/schemaconv/testdata/crds/openapiv2/container_v1beta1_containercluster.json
… pkg/schemaconv/testdata/crds/openapiv2/dataproc_v1beta1_dataprocworkflowtemplate.json
… pkg/schemaconv/testdata/crds/openapiv2/dlp_v1beta1_dlpdeidentifytemplate.json
… pkg/schemaconv/testdata/crds/openapiv2/monitoring_v1beta1_monitoringalertpolicy.json
… pkg/schemaconv/testdata/crds/openapiv2/monitoring_v1beta1_monitoringdashboard.json
… pkg/schemaconv/testdata/crds/openapiv2/privateca_v1beta1_privatecacertificate.json
… pkg/schemaconv/testdata/crds/openapiv2/privateca_v1beta1_privatecacertificateauthority.json
… pkg/schemaconv/testdata/crds/openapiv2/storagetransfer_v1beta1_storagetransferjob.json
… pkg/schemaconv/testdata/crds/openapiv2/vpcaccess_v1beta1_vpcaccessconnector.json
… pkg/schemaconv/testdata/crds/openapiv3/accesscontextmanager_v1beta1_accesscontextmanagerserviceperimeter.json
… pkg/schemaconv/testdata/crds/openapiv3/container_v1beta1_containercluster.json
… pkg/schemaconv/testdata/crds/openapiv3/dataproc_v1beta1_dataprocworkflowtemplate.json
… pkg/schemaconv/testdata/crds/openapiv3/dlp_v1beta1_dlpdeidentifytemplate.json
… pkg/schemaconv/testdata/crds/openapiv3/monitoring_v1beta1_monitoringalertpolicy.json
… pkg/schemaconv/testdata/crds/openapiv3/monitoring_v1beta1_monitoringdashboard.json
… pkg/schemaconv/testdata/crds/openapiv3/privateca_v1beta1_privatecacertificate.json
… pkg/schemaconv/testdata/crds/openapiv3/privateca_v1beta1_privatecacertificateauthority.json
… pkg/schemaconv/testdata/crds/openapiv3/storagetransfer_v1beta1_storagetransferjob.json
… pkg/schemaconv/testdata/crds/openapiv3/vpcaccess_v1beta1_vpcaccessconnector.json
… pkg/schemaconv/testdata/new-schema.yaml
… pkg/schemaconv/testdata/swagger.json