Intended to remove very expensive conversion via json & gnostic from apiserver.