@apelisse 2023-03-13
Reformatting benchmarks for clarity:
```
name                                                                                            old time/op    new time/op    delta
accesscontextmanager_v1beta1_accesscontextmanagerserviceperimeter.json-20    10.1ms ± 0%     0.1ms ± 0%  -99.20%  (p=0.008 n=5+5)
container_v1beta1_containercluster.json-20                                   10.8ms ± 0%     0.1ms ± 0%  -98.80%  (p=0.008 n=5+5)
dataproc_v1beta1_dataprocworkflowtemplate.json-20                            12.3ms ± 0%     0.1ms ± 0%  -98.87%  (p=0.008 n=5+5)
dlp_v1beta1_dlpdeidentifytemplate.json-20                                    18.2ms ± 0%     0.3ms ± 0%  -98.54%  (p=0.008 n=5+5)
monitoring_v1beta1_monitoringalertpolicy.json-20                             8.84ms ± 0%    0.04ms ± 0%  -99.52%  (p=0.008 n=5+5)
monitoring_v1beta1_monitoringdashboard.json-20                               26.0ms ± 0%     0.2ms ± 0%  -99.11%  (p=0.008 n=5+5)
privateca_v1beta1_privatecacertificate.json-20                               9.16ms ± 0%    0.08ms ± 0%  -99.14%  (p=0.008 n=5+5)
privateca_v1beta1_privatecacertificateauthority.json-20                      9.55ms ± 0%    0.09ms ± 0%  -99.09%  (p=0.008 n=5+5)
storagetransfer_v1beta1_storagetransferjob.json-20                           8.21ms ± 0%    0.05ms ± 0%  -99.39%  (p=0.008 n=5+5)
vpcaccess_v1beta1_vpcaccessconnector.json-20                                 7.50ms ± 0%    0.03ms ± 0%  -99.61%  (p=0.008 n=5+5)

name                                                                                            old alloc/op   new alloc/op   delta
accesscontextmanager_v1beta1_accesscontextmanagerserviceperimeter.json-20    6.15MB ± 0%    0.09MB ± 0%  -98.61%  (p=0.008 n=5+5)
container_v1beta1_containercluster.json-20                                   6.71MB ± 0%    0.12MB ± 0%  -98.21%  (p=0.008 n=5+5)
dataproc_v1beta1_dataprocworkflowtemplate.json-20                            7.91MB ± 0%    0.14MB ± 0%  -98.27%  (p=0.008 n=5+5)
dlp_v1beta1_dlpdeidentifytemplate.json-20                                    11.6MB ± 0%     0.3MB ± 0%  -97.65%  (p=0.008 n=5+5)
monitoring_v1beta1_monitoringalertpolicy.json-20                             5.59MB ± 0%    0.04MB ± 0%  -99.22%  (p=0.008 n=5+5)
monitoring_v1beta1_monitoringdashboard.json-20                               17.1MB ± 0%     0.3MB ± 0%  -98.52%  (p=0.008 n=5+5)
privateca_v1beta1_privatecacertificate.json-20                               5.81MB ± 0%    0.08MB ± 0%  -98.58%  (p=0.008 n=5+5)
privateca_v1beta1_privatecacertificateauthority.json-20                      5.97MB ± 0%    0.09MB ± 0%  -98.48%  (p=0.008 n=5+5)
storagetransfer_v1beta1_storagetransferjob.json-20                           5.24MB ± 0%    0.05MB ± 0%  -99.03%  (p=0.008 n=5+5)
vpcaccess_v1beta1_vpcaccessconnector.json-20                                 4.71MB ± 0%    0.03MB ± 0%  -99.40%  (p=0.008 n=5+5)

name                                                                                            old allocs/op  new allocs/op  delta
accesscontextmanager_v1beta1_accesscontextmanagerserviceperimeter.json-20     52.0k ± 0%      1.1k ± 0%  -97.88%  (p=0.008 n=5+5)
container_v1beta1_containercluster.json-20                                    60.5k ± 0%      1.7k ± 0%  -97.23%  (p=0.008 n=5+5)
dataproc_v1beta1_dataprocworkflowtemplate.json-20                             63.5k ± 0%      1.8k ± 0%  -97.17%  (p=0.008 n=5+5)
dlp_v1beta1_dlpdeidentifytemplate.json-20                                     85.8k ± 0%      3.3k ± 0%  -96.21%  (p=0.008 n=5+5)
monitoring_v1beta1_monitoringalertpolicy.json-20                              46.3k ± 0%      0.6k ± 0%  -98.65%  (p=0.008 n=5+5)
monitoring_v1beta1_monitoringdashboard.json-20                                99.5k ± 0%      3.1k ± 0%  -96.86%  (p=0.008 n=5+5)
privateca_v1beta1_privatecacertificate.json-20                                51.4k ± 0%      1.1k ± 0%  -97.85%  (p=0.008 n=5+5)
privateca_v1beta1_privatecacertificateauthority.json-20                       52.6k ± 0%      1.2k ± 0%  -97.75%  (p=0.008 n=5+5)
storagetransfer_v1beta1_storagetransferjob.json-20                            46.0k ± 0%      0.7k ± 0%  -98.45%  (p=0.008 n=5+5)
vpcaccess_v1beta1_vpcaccessconnector.json-20                                  41.7k ± 0%      0.4k ± 0%  -98.99%  (p=0.008 n=5+5)
```